RWA-1086 fixed code review comments
