+++ src/main/java/uk/gov/hmcts/reform/wataskmanagementapi/controllers/TaskActionsController.java
@@ -27,7 +27,6 @@ import uk.gov.hmcts.reform.wataskmanagementapi.controllers.advice.ErrorMessage;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.AssignTaskRequest;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.CompleteTaskRequest;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.NotesRequest;
-import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.TaskOperationRequest;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.response.GetTaskResponse;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.response.GetTaskRolePermissionsResponse;
 import uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.task.Task;
@@ -307,38 +306,6 @@ public class TaskActionsController extends BaseController {
             .ok()
             .cacheControl(CacheControl.noCache())
             .body(new GetTaskRolePermissionsResponse(taskRolePermissions));
-    }
-
-    @Operation(description = "performs specified operation like marking tasks to reconfigure and execute reconfigure.")
-    @ApiResponses({
-        @ApiResponse(responseCode = "204", description = "Task operation has been completed", content = {
-            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))}),
-        @ApiResponse(responseCode = "400", description = BAD_REQUEST),
-        @ApiResponse(responseCode = "403", description = FORBIDDEN),
-        @ApiResponse(responseCode = "401", description = UNAUTHORIZED),
-        @ApiResponse(responseCode = "415", description = UNSUPPORTED_MEDIA_TYPE),
-        @ApiResponse(responseCode = "500", description = INTERNAL_SERVER_ERROR)
-    })
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    @PostMapping(path = "/operation")
-    @SuppressWarnings("PMD.DataflowAnomalyAnalysis")
-    public ResponseEntity<Void> performOperation(@RequestHeader(SERVICE_AUTHORIZATION) String serviceAuthToken,
-                                                 @RequestBody TaskOperationRequest taskOperationRequest) {
-        boolean hasExclusiveAccessRequest =
-            clientAccessControlService.hasExclusiveAccess(serviceAuthToken);
-
-        if (hasExclusiveAccessRequest) {
-            taskManagementService.performOperation(
-                taskOperationRequest
-            );
-        } else {
-            throw new GenericForbiddenException(GENERIC_FORBIDDEN_ERROR);
-        }
-
-        return ResponseEntity
-            .noContent()
-            .cacheControl(CacheControl.noCache())
-            .build();
     }
 
     @ExceptionHandler(NoRoleAssignmentsFoundException.class)
